som_page: options with multiple selections specified in query string are not selected
Before c/s 8e58e9e80, we were asking jQuery to select each of an array of items, in a given select-multiple box.
In c/s 8e58e9e80, we mistakenly called decode() on this array, not realising it was an array. This caused jQuery to be unable to select the relevant options.
Instead, we should have called decode() on each element of the array.