Allow for floating-point x values
We were incorrectly calling parseInt() on each x-axis value in order that xs.sort() did a numeric sort. But now we are using a numeric_sort() function, so I think this is no longer necessary, and it is causing problems when the x values are floating-point (e.g. as on http://perf/?t=1521)