Pre-selecting options: append option if it doesn't already exist
Currently, if the URL query string contains an option that is not listed in a
particular select box, it is silently dropped. Instead, append this option to
the select box's options.

This scenario is common when one user creates a tiny URL from a page to which he
has manually added an option to one of the select boxes. This facility means
that users with whom he shares that tiny URL can view the same page.