Add % difference to the tooltips diff list
Add a percentage difference column to the list of differences shown in the
tooltip when genrating a tooltip with a previous selection. For numeric
values this is calculated as (newval - prevval)/prevval*100, and rounded
to 2 decimal places. If either value is non-numeric (eg. text fields) an
empty string is used. This is useful for quickly comparing eg. the y
results on a graph.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>