Add filtering to brief report
This commit adds a basic filtering capability to the brief report, based on the % of the right-most column (assumed to be the build under review). It can show/hide rows, based on a configurable (by a HTML input) threshold, the minimum % to consider. This defaults to 5%, which should reduce noise. We can also configure whether to show only regressions or also improvements.

Implementation:
- A new javascript ragebrief.js is included in static, included by the brief report HTML writer.
- The new inputs are added above the brief report table
- All inputs use immediate change handlers, we re-filter the table every time any of them change

Performance: The filtering is quite naive, we show all rows and then hide those that need hiding. I've tested this with a recent report with many rows/columns, and it's fast enough to be instant. If needed I can add some intelligence in future but for now this works well.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>