Add a preset to the graph page
When we analyse brief reports there are a series of steps we always take to produce a graph more useful than the generated graph link. This PR adds a button "Brief report analysis" which automatically performs these actions:

- Unselect build tag, select build date
- Select master branch
- Split by branch
- Select all build numbers
- Select all build tags
- Move the legend to bottom left
- Enable auto redraw
- Draw the graph

The building blocks are very simple functions `select`, `unselect`, `unselectAll`, which should make it trivial to build further presets in the future.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>