Automatically use brief report analysis options based on URL
Adds javascript to check the `location.hash`, so that if #brief_report_analysis is appended to the URL it will automatically apply the brief report preset options.
Modify graph links produced by the brief report to append #brief_report_analysis

Note: I used a # instead of an extra get parameter because the tiny URL ignores the hash - which makes sense because it's already set all the options in the preset so it doesn't need to remember it. It's also only used client-side.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>