Improve brief report filtering
Now instead of only looking at the last rows, it looks at all rows after the baseline (backward from the latest). If none of them have data it still hides the row, but if any of them have data then it uses that to assess the row even if it's not the very latest build. This should make the filter more useful when we only have partial data over multiple builds - we can put them all in and if any of them have a change in the last measurement we'll keep it visible.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>