Brief report: Add a freeze button to delete all hidden rows
Since filtering is dynamic we only show/hide rows when we change the values, we don't delete elements. However this means that if the report is saved then we like 1000+ rows in the HTML document and it's inefficient to parse.

This commit adds a 'Freeze' button to the filtering UI which when clicked (after a warning) permanently deletes the hidden rows, and disables further filtering (any results would be potentially incorrect because data has been deleted from the page). Now if the page is saved then only the visible rows are in the HTML and it can be parsed much more efficiently.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>