Add report quality metrics to the brief report
This commit adds two metrics to the brief report: Rows with data in last column, and rows with data in 2nd-to-last, but not last. This is a good indication of the quality of the data in the build being analysed (the last one in the table) against previous builds - if there are many rows with data in the previous build but not the last then tests are probably still ongoing.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>