Update to OCaml 4.08.1, Core v0.13, and Dune build system
Polymorphic comparison is not visible by default now, the default
comparison operators are for int.
Have to locally open the module for the appropriate type, or use
Polymorphic_compare.(a = b) if none is available.

Similarly all List lookup functions take an ~equal parameter now.

List.sort and List.dedup_and_sort take a ~compare parameter
(there is no dedup without sorting anymore).

Be careful with comparing NULL values when deduping and sorting
(NULL) fails the int_of_string which could cause List.dedup_and_sort to loose 1 value
as it considered it identical to NULL if the `cmp` function was used for
comparison. Make sure to use the String.compare function as the original
could would implicitly use.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>