Async postgres queries
Use a Deferred monad for performing queries.
This is similar to Lwt: instead of getting a response we get a
promise/future that we can evaluate later when the answer is available.

This'll allow us to do things like List.map and run bunch of queries in
parallel.

I simply run the postgresql query in a different thread on its own
connection (limited to number of cores/machine).

It would be possible to avoid using multiple threads and use
multiplexing via epoll and non-blocking reads/writes but libpq is
difficult to use: it will reconnect and change the file descriptor
behing your back which will cause epoll to fail because it is no longer
registered for epoll.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>