Show more accurate median statistics
The median is only the n/2 indexed element when
there are an odd number of elements, otherwise have to interpolate.

Also calculate the Tukey Fences, defined as Q1-1.5*IQR, Q3+1.5*IQR with IQR=Q3-Q1.
Useful for spotting outliers, or non-normal distributions.

We usually have a small number of data points, and no guarantees that
the data is normally distributed, so using median and quantile
statistics is probably the right choice since they are more robust to
outliers.
(e.g. due to caching effects the 1st data point in a VM boot time
 could be significantly larger than the rest, although still valid)

Signed-off-by: Edwin Török <edvin.torok@citrix.com>